+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to muglock are documented in this file.
 
-## [Unreleased]
+## [0.2.0] - 2026-08-18
 
 ### Changed